New title and desc, how to cite
Changed page title to just “Stingray” and changed the project
description, removed the h3 with Stingray and just have the project
description under the logo. Made each sentence a new line for each of
the paragraphs. Rephrased to remove GSoC 2017 (*tear*) from how to get
involved, but kept link to the ideas page since it’s still a good
starting point. Put the h4 on the same line as the title it’s using.
Added section on citing stingray with ASCL button and link. Added
Timelab shoutout under GSoC 2016 in acknowledgements.