Remove most content.
There was a lot of material duplicated between this page and the Stingray
documentation. This was doubly confusing, because that information was (in
genereal) only relevant to Stingray itself, and not to HENDRICS and DAVE which
are also covered by this page.

This PR makes this page into just a “landing page” for the StingraySoftware
organization. It provides links to all the associated projects; their pages
are assumed to provide all the information that readers will need, so it isn't
duplicated here.

(The one thing that wasn't already in the Stingray docs was the list of talks,
so I'm going to PR an addition of that there.)